Adding a basic discord bot to the repo
Here is a basic discord bot I wrote.  It's more of a template than anything else.  I'm planning on updating it with more as I un-stub the methods and add functionalities that I have implemented in other bots in the past.  This one has a rather unique command input structure.